+++ CHANGELOG.md
@@ -10,6 +10,7 @@ Changelog
 - `Group`: Add method `commonTraits`
 - Property `savedFileDisk` can be `null` and is `null` by default
 - Add option `showTemporaryFileWarning` (temporary file warning is no longer shown by default)
+- Exception `MediaCollectionNotRegisteredException` when media collection is not registered
 
 
 1.1.1 (2025-12-11)